+++ test.sh
@@ -4,5 +4,5 @@ rm .coverage
 poetry run black --check sortviz tests
 poetry run isort --check sortviz tests
 poetry run pytype .
-poetry run pytest
+poetry run pytest --cov=sortviz --cov-report term-missing
 poetry run coverage-badge -o svg/coverage.svg -f